The installer was 2,272 lines, and the scan that guards it stopped at one directory
`artifact_install.rs` was the last file in the desktop shell over DES-09's
600 lines. It is a directory now, split along the install's own stages --
read the manifest, download an artifact, expand it, check the disk it is
going on, stage the release, record what ended up installed -- with the
guards beside them in `tests/`, grouped the same way.

The interesting part is the guard that failed while doing it.

`a_release_build_ignores_every_runtime_redirecting_env_var` asserts that
three environment variables which choose *which* runtime a signed,
notarized build fetches are read only through `dev_override`. It scans
the whole shell for them, with one exception cut out by name. Two things
about it broke here, and both are the same mistake in different clothes:

`shell_source()` read `src/*.rs` and did not recurse, so the moment
`artifact_install` became a directory, 2,272 lines left the scan. The
walk is recursive now, and skips the guards themselves so a scan cannot
find its own needles.

The exception was cut as the text between two function names, which
assumed they stayed adjacent. They are in different files now, in the
other order. It cuts the one function out by body instead.

The first of those would have gone unnoticed: a security guard that
searches less of the tree still passes. It failed only because the cut
happened to panic. That is the second time in this branch a
source-scanning guard quietly shrank when the code moved, so
`shell_source()` now asserts it is reading more than twenty files.
